refactor(reservation): 예약 가능한 시간 조회 API 책임을 reservation이 갖도록 변경
