+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -6,10 +6,8 @@ import org.springframework.web.bind.annotation.*;
 import roomescape.service.ReservationTimeService;
 import roomescape.service.dto.request.ReservationTimeCreateRequest;
 import roomescape.service.dto.response.ReservationTimeResponse;
-import roomescape.service.dto.response.ReservationTimeStatusResponse;
 
 import java.net.URI;
-import java.time.LocalDate;
 import java.util.List;
 
 @RestController
@@ -22,15 +20,6 @@ public class ReservationTimeController {
     @GetMapping
     public ResponseEntity<List<ReservationTimeResponse>> getTimes() {
         final List<ReservationTimeResponse> results = reservationTimeService.getTimes();
-        return ResponseEntity.ok(results);
-    }
-
-    @GetMapping(params = {"date", "themeId"})
-    public ResponseEntity<List<ReservationTimeStatusResponse>> getTimesWithStatus(
-            @RequestParam(value = "date", required = false) LocalDate date,
-            @RequestParam(value = "themeId", required = false) Long themeId
-    ) {
-        final List<ReservationTimeStatusResponse> results = reservationTimeService.getTimesWithStatus(date, themeId);
         return ResponseEntity.ok(results);
     }
